Periodisering hører hjemme i FakturaserieGenerator
FakturaGenerator må vite om dagens dato og periodisering for å bestemme om fakturering havne i ny linje eller i en ny faktura.
